What is going on with darktable CI? These should have been caught a year ago.

* fix-clang-build:
  Follow-up, don't pass arguments to functions not taking arguments.
  Fix *NASTY* "passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x"
  Fix "floating-point comparison is always false; constant cannot be represented exactly in type 'float" clang diag
  Fix "variable '<...>' set but not used" clang diagnostics
  Fix "implicit truncation from 'int' to a one-bit wide bit-field changes value from 1 to -1" clang diag

+++ src/iop/graduatednd.c
@@ -236,7 +236,7 @@ static int set_grad_from_points(struct dt_iop_module_t *self, float xa, float ya
     if(r1 * r2 < 0) break;
   } while(v2 <= M_PI);
 
-  if(v2 == M_PI) return 9;
+  if(v2 == (float)M_PI) return 9;
 
   // set precision for the iterative check
   const float eps = .0001f;
